requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pandas==2.1.4
matplotlib==3.8.2
pytest==7.4.3
requests==2.31.0
sqlalchemy==2.0.23
